completed pdf writer, need to test sizes still
